requirements.txt
streamlit>=1.30.0

pandas>=2.0.0
numpy>=1.24.0

matplotlib>=3.7.0

statsmodels>=0.14.0

websocket-client>=1.6.0

# SQLite comes built-in with Python (no install needed)